+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Mapbox welcomes participation and contributions from everyone.
 * Add annotations drag handlers callbacks `dragBeginHandler`, `dragChangeHandler`, `dragEndHandler` to all Annotation types.
 * Add `allowHistTesting` modifier on `MapViewAnnotation`.
 * Fix taps propagation on `ViewAnnotation`.
+* Fix view annotations positioning on `.ignoresSafeArea(.all)` in SwiftUI
 * Bump core maps version to 11.2.0 and common sdk to 24.2.0.
 * Added Attribution and Telemetry pop-up dialogs and compass view content description translations for Arabic, Belarusian, Bulgarian, Catalan, Chinese Simplified, Chinese Traditional, Czech, Danish, Dutch, French, Galician, German, Hebrew, Italian, Japanese, Korean, Lithuanian, Norwegian, Polish, Belarusian, Russian, Spanish, Swedish, Ukranian and Vietnamese.